api/requirements.txt
Django
python-dotenv
djangorestframework
django-cors-headers
drf_yasg
torchvision
transformers
kagglehub
scikit-learn
faiss-cpu
